vm-core renamed to chain-core
