+++ framework/base/src/api.rs
@@ -30,4 +30,4 @@ pub use storage_api::*;
 pub use vm_api::VMApi;
 
 // Backwards compatibility.
-pub use crate::vm_core::builtin_func_names::*;
+pub use crate::chain_core::builtin_func_names::*;
