+++ framework/base/src/types.rs
@@ -16,7 +16,7 @@ pub use static_buffer::*;
 #[cfg(feature = "alloc")]
 pub use heap::*;
 
-pub use crate::vm_core::types::CodeMetadata;
-pub use crate::vm_core::types::EsdtLocalRole;
-pub use crate::vm_core::types::EsdtLocalRoleFlags;
-pub use crate::vm_core::types::EsdtTokenType;
+pub use crate::chain_core::types::CodeMetadata;
+pub use crate::chain_core::types::EsdtLocalRole;
+pub use crate::chain_core::types::EsdtLocalRoleFlags;
+pub use crate::chain_core::types::EsdtTokenType;
