+++ tools/rust-debugger/pretty-printers/multiversx_sc_lldb_pretty_printers.py
@@ -45,7 +45,7 @@ MANAGED_VEC_TYPE = f"{MOD_PATH}::managed_vec::ManagedVec<{DEBUG_API_TYPE}, {ANY_
 # 4. SC wasm - Managed multi value types
 
 # 5. VM core types
-MOD_PATH = "multiversx_chain_vm_core::types"
+MOD_PATH = "multiversx_chain_core::types"
 
 HEAP_ADDRESS_TYPE = f"{MOD_PATH}::address::Address"
 BOXED_BYTES_TYPE = f"{MOD_PATH}::boxed_bytes::BoxedBytes"
