+++ vm/Cargo.toml
@@ -34,7 +34,7 @@ rand = { version= "0.8.5", optional = true }
 rand_seeder = "0.3.0"
 ed25519-dalek = "2.1.0"
 
-[dependencies.multiversx-chain-vm-core]
+[dependencies.multiversx-chain-core]
 version = "=0.10.0"
 path = "../vm-core"
