+++ vm/src/tx_execution/builtin_function_mocks/esdt_nft/esdt_local_burn.rs
@@ -1,9 +1,9 @@
 use num_bigint::BigUint;
 
 use crate::{
+    chain_core::builtin_func_names::ESDT_LOCAL_BURN_FUNC_NAME,
     tx_execution::BlockchainVMRef,
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxLog, TxResult},
-    vm_core::builtin_func_names::ESDT_LOCAL_BURN_FUNC_NAME,
 };
 
 use super::super::builtin_func_trait::BuiltinFunction;
