+++ vm/src/tx_execution/builtin_function_mocks/esdt_nft/esdt_local_mint.rs
@@ -1,9 +1,9 @@
 use num_bigint::BigUint;
 
 use crate::{
+    chain_core::builtin_func_names::ESDT_LOCAL_MINT_FUNC_NAME,
     tx_execution::BlockchainVMRef,
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxLog, TxResult},
-    vm_core::builtin_func_names::ESDT_LOCAL_MINT_FUNC_NAME,
     world_mock::EsdtInstanceMetadata,
 };
