+++ vm/src/tx_execution/builtin_function_mocks/esdt_nft/esdt_nft_burn_mock.rs
@@ -1,10 +1,10 @@
 use num_bigint::BigUint;
 
 use crate::{
+    chain_core::builtin_func_names::ESDT_NFT_BURN_FUNC_NAME,
     tx_execution::BlockchainVMRef,
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxLog, TxResult},
     types::{top_decode_u64, top_encode_u64},
-    vm_core::builtin_func_names::ESDT_NFT_BURN_FUNC_NAME,
 };
 
 use super::super::builtin_func_trait::BuiltinFunction;
