+++ vm/src/tx_execution/builtin_function_mocks/general/change_owner_mock.rs
@@ -1,8 +1,8 @@
 use crate::{
+    chain_core::builtin_func_names::CHANGE_OWNER_BUILTIN_FUNC_NAME,
     tx_execution::BlockchainVMRef,
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxResult},
     types::VMAddress,
-    vm_core::builtin_func_names::CHANGE_OWNER_BUILTIN_FUNC_NAME,
 };
 
 use super::super::builtin_func_trait::BuiltinFunction;
