+++ vm/src/tx_execution/builtin_function_mocks/general/claim_developer_rewards_mock.rs
@@ -2,9 +2,9 @@ use num_bigint::BigUint;
 use num_traits::Zero;
 
 use crate::{
+    chain_core::builtin_func_names::CLAIM_DEVELOPER_REWARDS_FUNC_NAME,
     tx_execution::BlockchainVMRef,
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxResult},
-    vm_core::builtin_func_names::CLAIM_DEVELOPER_REWARDS_FUNC_NAME,
 };
 
 use super::super::builtin_func_trait::BuiltinFunction;
