+++ vm/src/tx_execution/builtin_function_mocks/general/delete_username_mock.rs
@@ -1,7 +1,7 @@
 use crate::{
+    chain_core::builtin_func_names::DELETE_USERNAME_FUNC_NAME,
     tx_execution::BlockchainVMRef,
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxResult},
-    vm_core::builtin_func_names::DELETE_USERNAME_FUNC_NAME,
 };
 
 use super::super::builtin_func_trait::BuiltinFunction;
