+++ vm/src/tx_execution/builtin_function_mocks/transfer/esdt_transfer_mock.rs
@@ -1,7 +1,7 @@
 use crate::{
+    chain_core::builtin_func_names::ESDT_TRANSFER_FUNC_NAME,
     tx_execution::{BlockchainVMRef, BuiltinFunctionEsdtTransferInfo},
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxLog, TxResult},
-    vm_core::builtin_func_names::ESDT_TRANSFER_FUNC_NAME,
 };
 
 use super::{
