+++ vm/src/vm_hooks/vh_handler/vh_blockchain.rs
@@ -1,6 +1,6 @@
 use crate::{
+    chain_core::builtin_func_names::*,
     types::{EsdtLocalRole, EsdtLocalRoleFlags, RawHandle, VMAddress},
-    vm_core::builtin_func_names::*,
     vm_hooks::VMHooksHandlerSource,
     world_mock::{EsdtData, EsdtInstance},
 };
