+++ vm/src/vm_hooks/vh_handler/vh_send.rs
@@ -1,10 +1,10 @@
 use crate::{
-    tx_mock::{AsyncCallTxData, Promise, TxFunctionName, TxTokenTransfer},
-    types::{top_encode_big_uint, top_encode_u64, RawHandle, VMAddress, VMCodeMetadata},
-    vm_core::builtin_func_names::{
+    chain_core::builtin_func_names::{
         ESDT_MULTI_TRANSFER_FUNC_NAME, ESDT_NFT_TRANSFER_FUNC_NAME, ESDT_TRANSFER_FUNC_NAME,
         UPGRADE_CONTRACT_FUNC_NAME,
     },
+    tx_mock::{AsyncCallTxData, Promise, TxFunctionName, TxTokenTransfer},
+    types::{top_encode_big_uint, top_encode_u64, RawHandle, VMAddress, VMCodeMetadata},
     vm_hooks::VMHooksHandlerSource,
 };
 use num_traits::Zero;
